feat: refactoring create_vars_file to modular structure
