+++ app/v1/modules/kubernetes_manager/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from .script_create_vars_file import ScriptCreateVarsFile
+
+
+__all__ = [
+    "ScriptCreateVarsFile",
+]
